fix(shared): drop emit_terminal_event_fn/citation_tools injection in execute_tool_call
Terminal events and citation sources are part of the expected execute_tool_call behavior. Keep those helpers colocated with tool execution so direct callers and tests do not need to pass extra injection arguments.
