+++ src/owui_ext/tools/sub_agent.py
@@ -45,8 +45,6 @@ from owui_ext.shared.prompt_utils import (
     _append_tool_server_prompts,
     merge_prompt_sections,
 )
-from owui_ext.shared.terminal_events import emit_terminal_tool_event
-from owui_ext.shared.tool_event_metadata import CITATION_TOOLS
 from owui_ext.shared.tool_execution import (
     execute_direct_tool_call,
     execute_tool_call,
@@ -774,8 +772,6 @@ async def run_sub_agent_loop(
                         "__messages__": current_messages,
                     },
                     event_emitter=event_emitter,
-                    emit_terminal_event_fn=emit_terminal_tool_event,
-                    citation_tools=CITATION_TOOLS,
                 )
 
                 # Emit status with tool result
