+++ tools/llm_review.py
@@ -208,70 +208,17 @@ def _append_tool_server_prompts(form_data: dict, extra_params: dict) -> dict:
     form_data["messages"] = messages
     return form_data
 
-# --- inlined from src/owui_ext/shared/terminal_events.py (owui_ext.shared.terminal_events) ---
+# --- inlined from src/owui_ext/shared/tool_execution.py (owui_ext.shared.tool_execution) ---
+import ast
 import json
 import logging
+import uuid
 from typing import Any, Callable, Optional
-_terminal_events_log = logging.getLogger("owui_ext.shared.terminal_events")
+from fastapi import Request
+_tool_execution_log = logging.getLogger("owui_ext.shared.tool_execution")
+_core_process_tool_result = None
 
 
-async def emit_terminal_tool_event(
-    *,
-    tool_function_name: str,
-    tool_function_params: dict,
-    tool_result: Any,
-    event_emitter: Optional[Callable],
-) -> None:
-    """Emit ``terminal:*`` UI events for Open Terminal tool results.
-
-    The function recognises only the names listed in
-    ``TERMINAL_EVENT_TOOLS`` (display_file / write_file /
-    replace_file_content / run_command) -- unknown names hit the
-    final ``else`` and silently return.
-    """
-    if not event_emitter:
-        return
-    if tool_function_name == "display_file":
-        path = (
-            tool_function_params.get("path", "")
-            if isinstance(tool_function_params, dict)
-            else ""
-        )
-        if not isinstance(path, str) or not path:
-            return
-        parsed = tool_result
-        if isinstance(parsed, str):
-            try:
-                parsed = json.loads(parsed)
-            except Exception:
-                parsed = tool_result
-        if isinstance(parsed, dict) and parsed.get("exists") is False:
-            return
-        event = {"type": "terminal:display_file", "data": {"path": path}}
-    elif tool_function_name in {"write_file", "replace_file_content"}:
-        path = (
-            tool_function_params.get("path", "")
-            if isinstance(tool_function_params, dict)
-            else ""
-        )
-        if not isinstance(path, str) or not path:
-            return
-        event = {
-            "type": f"terminal:{tool_function_name}",
-            "data": {"path": path},
-        }
-    elif tool_function_name == "run_command":
-        event = {"type": "terminal:run_command", "data": {}}
-    else:
-        return
-    try:
-        await event_emitter(event)
-    except Exception as exc:
-        _terminal_events_log.warning(
-            f"Error emitting terminal event for {tool_function_name}: {exc}"
-        )
-
-# --- inlined from src/owui_ext/shared/tool_event_metadata.py (owui_ext.shared.tool_event_metadata) ---
 CITATION_TOOLS: set[str] = {
     "search_web",
     "view_file",
@@ -287,16 +234,6 @@ TERMINAL_EVENT_TOOLS: set[str] = {
     "replace_file_content",
     "run_command",
 }
-
-# --- inlined from src/owui_ext/shared/tool_execution.py (owui_ext.shared.tool_execution) ---
-import ast
-import json
-import logging
-import uuid
-from typing import Any, Callable, Optional
-from fastapi import Request
-_tool_execution_log = logging.getLogger("owui_ext.shared.tool_execution")
-_core_process_tool_result = None
 
 
 async def _maybe_await(value: Any) -> Any:
@@ -421,23 +358,69 @@ def normalize_terminal_tools_result(
     return {}
 
 
+async def emit_terminal_tool_event(
+    *,
+    tool_function_name: str,
+    tool_function_params: dict,
+    tool_result: Any,
+    event_emitter: Optional[Callable],
+) -> None:
+    """Emit ``terminal:*`` UI events for Open Terminal tool results.
+
+    Recognises only the names listed in ``TERMINAL_EVENT_TOOLS``
+    (display_file / write_file / replace_file_content / run_command);
+    unknown names fall through silently.
+    """
+    if not event_emitter:
+        return
+    if tool_function_name == "display_file":
+        path = (
+            tool_function_params.get("path", "")
+            if isinstance(tool_function_params, dict)
+            else ""
+        )
+        if not isinstance(path, str) or not path:
+            return
+        parsed = tool_result
+        if isinstance(parsed, str):
+            try:
+                parsed = json.loads(parsed)
+            except Exception:
+                parsed = tool_result
+        if isinstance(parsed, dict) and parsed.get("exists") is False:
+            return
+        event = {"type": "terminal:display_file", "data": {"path": path}}
+    elif tool_function_name in {"write_file", "replace_file_content"}:
+        path = (
+            tool_function_params.get("path", "")
+            if isinstance(tool_function_params, dict)
+            else ""
+        )
+        if not isinstance(path, str) or not path:
+            return
+        event = {
+            "type": f"terminal:{tool_function_name}",
+            "data": {"path": path},
+        }
+    elif tool_function_name == "run_command":
+        event = {"type": "terminal:run_command", "data": {}}
+    else:
+        return
+    try:
+        await event_emitter(event)
+    except Exception as exc:
+        _tool_execution_log.warning(
+            f"Error emitting terminal event for {tool_function_name}: {exc}"
+        )
+
+
 async def execute_tool_call(
     tool_call: dict,
     tools_dict: dict,
     extra_params: dict,
     event_emitter: Optional[Callable] = None,
-    *,
-    emit_terminal_event_fn: Optional[Callable] = None,
-    citation_tools: Optional[set] = None,
 ) -> dict:
-    """Execute a single tool call and return ``{tool_call_id, content}``.
-
-    ``emit_terminal_event_fn`` and ``citation_tools`` are injected by the
-    caller (typically ``shared.terminal_events.emit_terminal_tool_event``
-    and ``shared.tool_event_metadata.CITATION_TOOLS``). They cannot be
-    imported here directly because the inliner forbids a shared dep
-    module from mixing external imports with cross-shared imports.
-    """
+    """Execute a single tool call and return ``{tool_call_id, content}``."""
     if not isinstance(tool_call, dict):
         return {
             "tool_call_id": str(uuid.uuid4()),
@@ -535,13 +518,12 @@ async def execute_tool_call(
         tool_result = f"Tool '{tool_function_name}' not found"
 
     if emit_terminal_event:
-        if emit_terminal_event_fn is not None:
-            await emit_terminal_event_fn(
-                tool_function_name=tool_function_name,
-                tool_function_params=tool_function_params,
-                tool_result=tool_result,
-                event_emitter=event_emitter,
-            )
+        await emit_terminal_tool_event(
+            tool_function_name=tool_function_name,
+            tool_function_params=tool_function_params,
+            tool_result=tool_result,
+            event_emitter=event_emitter,
+        )
         if event_emitter and tool_result_files:
             await event_emitter({"type": "files", "data": {"files": tool_result_files}})
         if event_emitter and tool_result_embeds:
@@ -555,12 +537,7 @@ async def execute_tool_call(
         except Exception:
             tool_result = str(tool_result)
 
-    if (
-        event_emitter
-        and tool_result
-        and citation_tools
-        and tool_function_name in citation_tools
-    ):
+    if event_emitter and tool_result and tool_function_name in CITATION_TOOLS:
         try:
             from open_webui.utils.middleware import get_citation_source_from_tool_result
 
@@ -4199,8 +4176,6 @@ async def run_agent_loop(
                         "__messages__": current_messages,
                     },
                     event_emitter=event_emitter,
-                    emit_terminal_event_fn=emit_terminal_tool_event,
-                    citation_tools=CITATION_TOOLS,
                 )
 
                 # Emit status with tool result preview
@@ -4404,8 +4379,6 @@ async def run_agent_loop(
                                             "__messages__": current_messages,
                                         },
                                         event_emitter=event_emitter,
-                                        emit_terminal_event_fn=emit_terminal_tool_event,
-                                        citation_tools=CITATION_TOOLS,
                                     )
                                 except Exception as exec_exc:
                                     log.warning(
